requirements.txt
DateTime==4.3
flake8==3.7.7
Flask==1.0.2
PyYAML==5.3
termcolor==1.1.0
requests==2.24.0
sockets==1.0.0
Flask-SocketIO==5.0.1
